requirements.txt
openai-whisper>=20231117
stable-ts>=2.16.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
language-tool-python>=2.8
textgrid>=1.5
rich>=13.0.0
pydub>=0.25.1
pyannote.audio>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.31.0
flask-cors>=6.0.0

# ML Grammar Checker (optional, but recommended)
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
